requirements.txt
fastapi
uvicorn
pydantic
jinja2
python-multipart
emails  # 用于发送邮件
requests  # 用于 HTTP 请求
beautifulsoup4  # 用于解析 HTML
lxml  # 用于解析 XML sitemap
python-dotenv
pytest 
httpx  # 用于异步 HTTP 请求（Firecrawl MCP）
apscheduler  # 用于定时任务
redis  # Redis 向量数据库
openai  # OpenAI API
langchain  # 大模型框架
langchain-community  # LangChain 社区包（Firecrawl Loader）
json-repair  # 用于修复损坏的JSON
nest-asyncio  # 允许嵌套事件循环
pyyaml  # YAML配置文件支持
ruamel.yaml  # 保留注释的YAML处理